WIP Restructure - Started creating "DIR"class, which has the Run() function.
Run() can run both a file and a string (for REPL). Started to incorporate this into the Lexer class and then Parser class. Made a decision to prefer encapsulation over inheritance (Lexer and Parser are in DIR instead of dealing with the Module class). I might use Module for error handling code, but still figuring that out.
Added new Error class.